002-change customerId type to integer in the entities.

+++ loan/src/main/java/huseyin/ocal/loan/entity/Loan.java
@@ -20,7 +20,7 @@ public class Loan {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private int loanNumber;
 
-    private Long customerId;
+    private int customerId;
 
     @CreationTimestamp
     @Column(updatable = false)
